Adding permissions

+++ server/api/users/[id].delete.js
@@ -1,7 +1,22 @@
 import User from '~~/server/models/user';
+import AuthHelper from '~~/server/helpers/auth.helper';
 
+const { checkDefaultAdmin, verifyToken } = AuthHelper;
+
+/**
+ * Deletes a user
+ */
 export default defineEventHandler(async (event) => {
   try {
+    const { authorization } = event.node.req.headers;
+    const userObj = verifyToken(authorization);
+    if (!checkDefaultAdmin(userObj)) {
+      return {
+        status: 403,
+        message: 'You do not have permision',
+        data: null
+      };
+    }
     const id = event.context.params.id;
     await User.findByIdAndDelete(id);
     return {
@@ -12,7 +27,7 @@ export default defineEventHandler(async (event) => {
   } catch (e) {
     return {
       status: 400,
-      message: 'Something went horribly wrong',
+      message: e.message,
       data: null
     };
   }
